Max code clean
