+++ BeocreateConnect/main.ts
@@ -19,111 +19,22 @@ SOFTWARE.
 */
 
 // BEOCREATE CONNECT
-import { app, Menu, BrowserWindow, ipcMain, nativeTheme, systemPreferences, shell, MenuItem } from 'electron';
-
-import windowStateKeeper from 'electron-window-state';
+import { app, BrowserWindow, ipcMain, nativeTheme, systemPreferences } from 'electron';
 
 import { Browser, tcp } from "dnssd2";
-import { networkInterfaces } from "os";
 
 import { NodeSSH } from 'node-ssh';
 import { fetch } from 'cross-fetch';
 import { ClientChannel } from 'ssh2';
-
-import defineMenu from './utils';
+import { createWindow, refreshProducts, setProductInfo } from './utils';
 
 let debug = false;
-let activeWindow: boolean | null = true;
-
-let win: BrowserWindow | null = null;
+export let win: BrowserWindow | null = null;
   
-function createWindow () {
-	const mainWindowState = windowStateKeeper({
-	    defaultWidth: 820,
-	    defaultHeight: 600
-	});
-
-	win = new BrowserWindow({
-		x: mainWindowState.x,
-		y: mainWindowState.y,
-		width: mainWindowState.width,
-		height: mainWindowState.height, 
-		minWidth: 450, 
-		minHeight: 300, 
-		titleBarStyle: "hiddenInset", 
-		title: "Beocreate Connect",
-		frame: false,
-		show: false,
-		backgroundColor: '#FFFFFF', 
-		webPreferences: { experimentalFeatures: false, nodeIntegration: true, contextIsolation: false,}
-	});
-
-	defineMenu(win);
-	
-	mainWindowState.manage(win);
-  
-	// @todo : use build/index.html for release bundling.
-	// demistify that
-    //win.loadFile('build/index.html')
-    win.loadFile('index.html');
-
-	// win.webContents.openDevTools()
-
-    win.webContents.on('did-finish-load', () => {
-		if (process.platform == 'darwin') {
-			win?.webContents.send('colourSchemeIsDark', nativeTheme.shouldUseDarkColors);
-		}
-		win?.webContents.send('styleForWindows', process.platform !== 'darwin');
-		setTimeout(function() {
-			win?.show();
-			setTimeout(function() {
-				startDiscovery();
-				startCheckingIPAddress();
-				startManualDiscovery();
-			}, 500);
-		}, 100);
-    });
-    
-    win.on('closed', () => {
-	  win = null;
-	  stopManualDiscovery();
-
-	  if(ipCheckInterval) {
-		clearInterval(ipCheckInterval);
-	  }
-
-	  stopDiscovery();
-	  if(sshInstance) {
-		sshInstance.dispose();
-	  }
-    });
-    
-    win.on('focus', () => {
-    	win?.webContents.send('windowEvent', "activate");
-		refreshProducts(null);
-		activeWindow = true;
-    });
-    
-    win.on('blur', () => {
-    	win?.webContents.send('windowEvent', "resignActive");
-		activeWindow = null;
-    });
-	
-	win.on("enter-full-screen", () => {
-		win?.webContents.send('windowEvent', "fullScreen");
-	});
-	
-	win.on("leave-full-screen", () => {
-		win?.webContents.send('windowEvent', "windowed");
-	});
-	
-	win.webContents.on('new-window', function(event, url){
-	  event.preventDefault();
-	  shell.openExternal(url);
-	});
-}
-  
-app.on('ready', createWindow);
+app.on('ready', () => {
+	win = createWindow();
+	onRefreshProducts();
+});
 
 app.on('window-all-closed', () => {
 	if (process.platform !== 'darwin') {
@@ -132,9 +43,9 @@ app.on('window-all-closed', () => {
 });
 
 app.on('activate', () => {
-	if (win === null) {
-		createWindow();
-	}
+	if(win) { return; }
+	
+	win = createWindow();
 });
 
 // DARK / LIGHT MODE
@@ -147,7 +58,7 @@ if (process.platform == "darwin" && win) {
 	)
 }
 
-interface Service {
+export interface Service {
 	host: string;
 	port: string;
 	name: string;
@@ -158,7 +69,7 @@ interface Service {
 }
   
 // FIND BEOCREATE SYSTEMS
-let browser: {
+export let browser: {
 	on: (eventName: string, callback: (service: Service) => void) => void;
 	start: () => void;
 	stop: () => void;
@@ -200,7 +111,7 @@ export function stopDiscovery() {
 	}
 }
 
-let products: Record<string, Service> = {};
+export let products: Record<string, Service> = {};
 let bonjourProductCount = 0;
 
 function discoveryEvent(event: string, service: Service): void {
@@ -227,122 +138,6 @@ function discoveryEvent(event: string, service: Service): void {
 	
 }
 
-function refreshProducts(services?: Service[] | null): void {
-	if (services == null) {
-		services = [];
-		if (browser) {
-			services = browser.list();
-		}
-	}
-
-	if (services.length == 0 && manuallyDiscoveredProduct) {
-		services.push(manuallyDiscoveredProduct);
-	}
-
-	if (services) {
-		// Find out which services have been added.
-		for (let s = 0; s < services.length; s++) {
-			if (!products[services[s].fullname]) {
-				setProductInfo(services[s]); // Adds product.
-				//console.log(products[services[s].fullname].addresses);
-				win?.webContents.send('addProduct', products[services[s].fullname]);
-			}
-		}
-		
-		// Find out which services have been removed.
-		for (let fullname in products) {
-			let serviceFound = -1;
-			for (let s = 0; s < services.length; s++) {
-				if (services[s].fullname == fullname) serviceFound = s;
-			}
-			if (serviceFound == -1) {
-				win?.webContents.send('removeProduct', products[fullname]);
-				delete products[fullname]; // Removes product.
-			}
-		}
-	}
-}
-
-interface Product {
-	fullname: string | number;
-	addresses: string[];
-	host: string;
-	port: string;
-	name: string;
-	modelID: string | null;
-	modelName: string | null;
-	productImage: string | null;
-	systemID: string | null;
-	systemStatus: string | null;
-	boundTo: string | undefined;
-	manual: boolean;
-	txt: Record<string, string>;
-}
-
-function setProductInfo(service: Service): Product {
-	let modelID = null;
-	let modelName = null;
-	let systemID = null;
-	let systemStatus = null;
-	let productImage = null;
-
-	for (const key in service.txt) {
-		if(!service.txt.hasOwnProperty(key)) {
-			continue;
-		}
-
-		switch (key) {
-			case "type":
-			case "device_type":
-				modelID = service.txt[key];
-				break;
-			case "typeui":
-				modelName = service.txt[key];
-				break;
-			case "id":
-			case "device_id":
-				systemID = service.txt[key];
-				break;
-			case "status":
-			case "device_status":
-				systemStatus = service.txt[key];
-				break;
-			case "image":
-				productImage = service.txt[key];
-				break;
-		}
-	}
-
-	let product: Product = {
-		fullname: service.fullname,
-		addresses: service.addresses,
-		host: service.host,
-		port: service.port,
-		name: service.name,
-		modelID: modelID,
-		modelName: modelName,
-		productImage: productImage,
-		systemID: systemID,
-		systemStatus: systemStatus,
-		boundTo: undefined,
-		manual: false,
-		txt: {}
-	};
-
-	// @note : auto-magic bind (jack button => reflect by bolt on UI)
-	if(currentRouting && currentRouting.from === service.name) {
-		product.boundTo = currentRouting.to;
-	} else if(currentRouting && currentRouting.to === service.name) {
-		product.boundTo = currentRouting.from;
-	}
-
-	if (service.manual) {
-		product.manual = true;
-	} 
-	products[service.fullname] = product;
-	return product;
-}
-
 ipcMain.on("getAllProducts", (event, arg) => {
 	win?.webContents.send('discoveredProducts', products);
 });
@@ -367,9 +162,9 @@ function findProduct(name: string) {
 	return productKey && products[productKey] ? products[productKey] : null;
 }
 
-let clientChannel: ClientChannel | null = null;
-let sshInstance: NodeSSH | null = null;
-let currentRouting: {
+export let clientChannel: ClientChannel | null = null;
+export let sshInstance: NodeSSH | null = null;
+export let currentRouting: {
 	from: string;
 	to: string;
 } | null = null;
@@ -465,7 +260,7 @@ ipcMain.on("bindAuxToAMP", (event, arg) => {
 	connectSSH();
 });
 
-let manuallyDiscoveredProduct: Service | null = null;
+export let manuallyDiscoveredProduct: Service | null = null;
 let manualDiscoveryInterval: NodeJS.Timer | null = null;
 let manualDiscoveryAddress = "10.0.0.1";
 
@@ -523,64 +318,4 @@ export function stopManualDiscovery(): void {
 	}
 
 	clearInterval(manualDiscoveryInterval);
-}
-
-let ipCheckInterval: NodeJS.Timer | null = null;
-function startCheckingIPAddress(): void {
-	hasIPChanged();
-	ipCheckInterval = setInterval(function() {
-		if (activeWindow) {
-			if (hasIPChanged()) {
-				startDiscovery();
-				startManualDiscovery();
-			}
-		}
-	}, 10000);
-}
-
-let oldIPs: string[] = [];
-function hasIPChanged(): boolean {
-	let ifaces = networkInterfaces();
-	let newIPs = []
-	for (let iface in ifaces) {
-		const currentInterface = ifaces[iface];
-		if(!currentInterface) { continue; }
-
-		for (let i = 0; i < currentInterface.length ?? 0; i++) {
-			if (currentInterface[i].family == "IPv4") {
-				newIPs.push(currentInterface[i].address);
-			}
-		}
-	}
-	if (equals(oldIPs, newIPs)) {
-		return false;
-	} else {
-		oldIPs = newIPs;
-		return true;
-	}
-}
-
-function equals(arrA: (unknown|unknown[])[], arrB: (unknown | unknown[])[]): boolean {
-    if (!arrA){
-		return false;
-	}
-
-    if (arrA.length != arrB?.length) {
-        return false;
-	}
-
-    for (let i = 0, l=arrA.length; i < l; i++) {
-		const item = arrA[i];
-		const itemB = arrB[i];
-
-		if(item instanceof Array &&  itemB instanceof Array) {
-			if(!equals(item, itemB)) {
-				return false;
-			}
-		} else if (item != itemB) {
-			return false;
-		}
-    }
-
-    return true;
 }
